requirements_web.txt
Flask==3.0.3
Flask-Login==0.6.3
requests>=2.28.0
Werkzeug>=3.0.3
Pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
openpyxl>=3.1.0
img2pdf>=0.5.0
